requirements.txt
fastapi
dotenv
SQLAlchemy
rich
aiohttp
google-adk
google-generativeai
